server/requirements.txt
# Cloud-Based Digital Library Backend Dependencies

# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0

# AWS SDK (only for Cognito)
boto3==1.34.0
botocore==1.34.0

# Environment variables
python-dotenv==1.0.0

# JWT / JOSE
python-jose[cryptography]==3.3.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0

# Production server
gunicorn==21.2.0

# Security
cryptography>=41.0.0

# Utilities
python-dateutil==2.8.2

# PDF processing for cover extraction
PyPDF2==3.0.1
Pillow==10.1.0
pdf2image==1.16.3
